[ADD] Top5TextLoop.jsx [UPDATE] home.jsx
