+++ components/app/home/index.js
@@ -1,0 +1,1 @@
+export { Top5TextLoop } from './top5TextLoop';
