+++ package.json
@@ -15,7 +15,8 @@
     "next": "11.1.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-icons": "^4.2.0"
+    "react-icons": "^4.2.0",
+    "react-text-loop": "^2.3.0"
   },
   "devDependencies": {
     "@emotion/babel-plugin": "^11.3.0",
