+++ pages/home.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { LayoutContainer, TypoGraphy } from 'components/common/index';
 import { customColor } from 'constants/index';
-import { RiArrowDropDownLine } from 'react-icons/ri';
+import { Top5TextLoop } from 'components/app/home';
 const home = () => {
   return (
     <LayoutContainer>
@@ -37,46 +37,7 @@ const home = () => {
             </TypoGraphy>
           </EnrollmentButton>
         </EnrollmentWrapper>
-        <Top5Section>
-          <Top5Title>
-            <TypoGraphy
-              type="h4"
-              color={customColor.white}
-              textAlign="left"
-              fontWeight="bold">
-              TOP 5
-            </TypoGraphy>
-          </Top5Title>
-          <Top5ItemWrapper>
-            <ItemRanking>
-              <TypoGraphy
-                type="body1"
-                color={customColor.skyBlue}
-                textAlign="center"
-                fontWeight="bold">
-                1
-              </TypoGraphy>
-            </ItemRanking>
-            <ItemTitle>
-              <TypoGraphy
-                type="body1"
-                color={customColor.black}
-                textAlign="left">
-                학교 폭파 해주세요. 빼애애애애애애액. 으아앙
-              </TypoGraphy>
-            </ItemTitle>
-            <ItemSubNumber>
-              <TypoGraphy
-                type="body1"
-                color={customColor.skyBlue}
-                textAlign="center"
-                fontWeight="bold">
-                8,888,888
-              </TypoGraphy>
-            </ItemSubNumber>
-            <RiArrowDropDownLine style={DropDownStyle} />
-          </Top5ItemWrapper>
-        </Top5Section>
+        <Top5TextLoop />
       </Container>
     </LayoutContainer>
   );
@@ -92,7 +53,7 @@ const Container = styled.div`
   margin-left: auto;
   margin-right: auto;
   padding: 16px;
-  background-color: ${customColor.blue};
+  background-color: ${customColor.deepBlue};
 `;
 
 const Title = styled.div`
@@ -121,7 +82,7 @@ const UrlTitle = styled.div`
   align-items: center;
   width: 11%;
   height: 100%;
-  border-right: 2px solid ${customColor.gray};
+  border-right: 2px solid ${customColor.grayBg};
 `;
 const InputUrl = styled.div`
   width: 70%;
@@ -134,45 +95,3 @@ const EnrollmentButton = styled.button`
   width: 13%;
   border-radius: 20px;
 `;
-
-const Top5Section = styled.div``;
-const Top5Title = styled.div`
-  margin-left: 20px;
-  margin-bottom: 5px;
-`;
-const Top5ItemWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 600px;
-  height: 45px;
-  padding: 0px 10px;
-
-  background-color: ${customColor.white};
-  border-radius: 20px;
-`;
-const ItemRanking = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 5%;
-`;
-const ItemTitle = styled.div`
-  display: flex;
-  align-items: center;
-  width: 75%;
-`;
-const ItemSubNumber = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 15%;
-`;
-const DropDownStyle = {
-  display: 'flex',
-  fontSize: '40px',
-  justifyContent: 'center',
-  alignItems: 'center',
-  cursor: 'pointer',
-  width: '5%',
-};
